Power down ADC voltage regulator
Power down the ADC voltage regulator as apart of the Adc::power_off()
method.

Additionally remove powering up the regulator in the Adc::new() method
to allow the Adc struct to be instantiated while in low power run mode.
The regulator will automatically be powered on when ADEN is set.

This is a partial fix for issue #231